+++ app/core/utils/progress_bar.rb
@@ -6,55 +6,32 @@ class ProgressBar
   attr_writer :limit_per_second
 
   def initialize(total)
-    @total = total
-    @lines_taken = 0
-    @semaphore = Mutex.new
-    @last_print_time = 0
-    @limit_per_second = 10
+    @total = total.to_f
+    @last_print_time = Time.now
+    @last_line_length = 0 # Initialize the last line length
+    @limit_per_second = 1 # Default to updating once per second
   end
 
   def print(progress, status, force: false)
-    return unless force || can_print?
-    status ||= ""
+    current_time = Time.now
+    # Calculate if we should update the progress bar based on the limit_per_second
+    should_print = force || (current_time - @last_print_time) > (1.0 / @limit_per_second)
+    return unless should_print
+    return unless PatchKitConfig.show_progress_bar_updates
 
-    return unless @semaphore.try_lock
+    # Update the progress display
+    percent = (progress.to_f / @total * 100).round(2)
+    progress_bar = "=" * (percent / 10).to_i + " " * (10 - percent / 10).to_i
+    line = "[#{progress_bar}] #{percent}% - #{status}"
 
-    console_width = IO.console ? IO.console.winsize[1] : 0
+    # Clear the last printed line by overwriting it with spaces if necessary
+    clear_line = " " * [@last_line_length, line.length].max
 
-    return if console_width < 6
-
-    if @lines_taken > 0
-      $stdout.write "\e[1A\r"
-      $stdout.write " " * @previous_status_length if @previous_status_length > status.length
-      $stdout.write "\e[#{@lines_taken - 1}A\r"
-      $stdout.flush
-    end
-
-    progress_bar_length = console_width - 5
-    progress_length = (progress_bar_length.to_f * get_progress_value(progress)).round
-    if progress_length == progress_bar_length
-      $stdout.write "|#{'=' * progress_length}|\n#{status}\n"
-    else
-      remaining_length = progress_bar_length - progress_length
-      $stdout.write "|#{'=' * [0, (progress_length - 1)].max}#{'>' * [1, (progress_length)].min}#{'-' * remaining_length}|\n#{status}\n"
-    end
+    $stdout.print "\r#{clear_line}\r" # Overwrite the old line with spaces, then return to the start of the line
+    $stdout.print line # Print the new progress bar and status
     $stdout.flush
-    @lines_taken = (status.length.to_f / console_width.to_f).ceil + 1
-    @previous_status_length = status.length
 
-    @last_print_time = Time.now.to_f
-  ensure
-    @semaphore.unlock if @semaphore.owned?
+    @last_print_time = current_time
+    @last_line_length = line.length # Remember the new line length for next update
   end
-
-  private
-
-    def can_print?
-      return false unless PatchKitConfig.show_progress_bar_updates
-      (Time.now.to_f - @last_print_time) >= 1.0 / @limit_per_second
-    end
-
-    def get_progress_value(progress)
-      [[progress.to_f / @total.to_f, 0].max, 1].min
-    end
 end